**Functional Hamburger Menu Nav Bar & Mobile or Desktop personalized navigation - first stage**
Implement the **context** to hold the state of whether or not is the hamburger menu active, sharing the state across the mobile nav, the main page to apply conditional rendering of the hamburger menu and the hamburger menu to deactivate it. Apply event listener the hamburger icon the activate or deactivate the menu.

Another update is the conditional rendering of the navigation top menu which based on the **window width** defines if an horizontal nav bar should be render or a mobile approach with hamburger menu. Use the resize event listener to keep it sensible to browser window resize and not only the initial width.